requirements.txt
Flask==2.3.3
langchain-core==1.2.5
langchain-openai==1.0.1
pinecone==6.0.2
pydantic==2.12.3
python-dotenv==1.0.0
tqdm==4.66.4